tests/conftest.py
from collections.abc import Iterator

import pytest
from sysetup.models import Path


def provision_path() -> Iterator[Path]:
    with Path.tempfile() as path:
        yield path
    assert not path.exists()


@pytest.fixture()
def path() -> Iterator[Path]:
    yield from provision_path()
